Refactor EligibilityResult
